feat: extract map workspace to connect to application
